Move internal container annotations to the appropriate packages
