+++ flow-server/src/main/java/com/vaadin/ui/common/InternalContainerAnnotationForJS.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.annotations;
+package com.vaadin.ui.common;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -21,27 +21,25 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import com.vaadin.ui.common.StyleSheet;
-
 /**
- * Annotation enabling using multiple {@link StyleSheet @StyleSheet}
+ * Annotation enabling using multiple {@link JavaScript @JavaScript}
  * annotations.
  * <p>
  * <b>NOT meant to be used</b>, for multiple style sheet dependencies,
- * {@link StyleSheet @StyleSheet} should be used instead.
+ * {@link JavaScript @JavaScript} should be used instead.
  *
  * @author Vaadin Ltd
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @Documented
-public @interface InternalContainerAnnotationForSS {
+public @interface InternalContainerAnnotationForJS {
 
     /**
-     * Not to be used, instead multiple {@link StyleSheet @StyleSheet}
+     * Not to be used, instead multiple {@link JavaScript @JavaScript}
      * annotations should be used.
      *
-     * @return an array of the style sheet annotations
+     * @return an array of the JavaScript annotations
      */
-    StyleSheet[] value();
+    JavaScript[] value();
 }
